Rearrange session management so we can use the right primitives
